fix: use ctx for response
